cambios de aportacionService a medias

+++ KiloApi/src/main/java/com/Triana/Salesinaos/KiloApi/model/DetalleAportacionPK.java
@@ -17,5 +17,5 @@ public class DetalleAportacionPK implements Serializable {
 
     private Long aportacionId;
 
-    private Long numLinea;
+    private int numLinea;
 }
